Create README.md
This commit includes the initial version of the README file for the project "Air Pollution and Time-Series Analysis in Delhi NCR Region Using Sentinel-5P Data". The README provides a comprehensive overview of the project, detailing the steps for analyzing air pollution concentrations (SO2, NO2, and Aerosols) over the Delhi NCR region. It includes sections on description, installation, usage, results, contributing, license, and contact information. The README serves as a guide for users to understand and replicate the analysis using Google Earth Engine.